requirements.txt
openai>=1.30.0
azure-identity>=1.17.0
azure-cognitiveservices-speech>=1.40.0
azure-storage-blob>=12.20.0
azure-storage-queue>=12.10.0
azure-data-tables>=12.5.0
beautifulsoup4>=4.12.0
requests>=2.31.0
click>=8.1.0
python-dotenv>=1.0.0
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
httpx>=0.27.0
